MR-2669 - renamed path
