+++ client/components/Modals/get-expert-advice-modal.vue
@@ -71,7 +71,7 @@
 <script>
 import { required, email, maxLength } from 'vuelidate/lib/validators';
 import { phone } from '@/helpers/validators';
-import {phoneHandler} from '@/helpers/mixins';
+import {phoneHandler} from '@/mixins/phoenHandler';
 import FormCheckboxes from '@/components/ui/form-checkboxes';
 import UIButton from '@/components/ui/UIButton';
