+++ client/components/Modals/get-your-trusted-it-partner-modal.vue
@@ -73,7 +73,7 @@ import { required, email, maxLength } from 'vuelidate/lib/validators';
 import { phone } from '@/helpers/validators';
 import FormCheckboxes from '@/components/ui/form-checkboxes';
 import UIButton from '@/components/ui/UIButton';
-import {phoneHandler} from '@/helpers/mixins';
+import {phoneHandler} from '@/mixins/phoenHandler';
 
 export default {
   name: 'get-your-trusted-it-partner',
