+++ client/components/Modals/teams-modal.vue
@@ -79,7 +79,7 @@
 <script>
 import { required, email, maxLength } from 'vuelidate/lib/validators';
 import { phone } from '@/helpers/validators';
-import {phoneHandler} from '@/helpers/mixins';
+import {phoneHandler} from '@/mixins/phoenHandler';
 import FormCheckboxes from '@/components/ui/form-checkboxes';
 import RadioList from '@/components/ui/radio-list';
 import UIButton from '@/components/ui/UIButton';
